Removed mostly all warnings and added assertion mechanism

+++ code/source/render/passes/albedo_pass.cpp
@@ -23,7 +23,7 @@ void gdr::albedo_pass::Initialize(void)
         
         {
           bindlessTexturesDesc[0].BaseShaderRegister = GDRGPUTexturePoolSlot;
-          bindlessTexturesDesc[0].NumDescriptors = Render->CreationParams.MaxTextureAmount;
+          bindlessTexturesDesc[0].NumDescriptors = (UINT)Render->CreationParams.MaxTextureAmount;
           bindlessTexturesDesc[0].OffsetInDescriptorsFromTableStart = 0;
           bindlessTexturesDesc[0].RangeType = D3D12_DESCRIPTOR_RANGE_TYPE_SRV;
           bindlessTexturesDesc[0].RegisterSpace = GDRGPUTexturePoolSpace;
@@ -33,7 +33,7 @@ void gdr::albedo_pass::Initialize(void)
 
         {
           cubeBindlessTexturesDesc[0].BaseShaderRegister = GDRGPUCubeTexturePoolSlot;
-          cubeBindlessTexturesDesc[0].NumDescriptors = Render->CreationParams.MaxTextureAmount;
+          cubeBindlessTexturesDesc[0].NumDescriptors = (UINT)Render->CreationParams.MaxTextureAmount;
           cubeBindlessTexturesDesc[0].OffsetInDescriptorsFromTableStart = 0;
           cubeBindlessTexturesDesc[0].RangeType = D3D12_DESCRIPTOR_RANGE_TYPE_SRV;
           cubeBindlessTexturesDesc[0].RegisterSpace = GDRGPUCubeTexturePoolSpace;
@@ -43,7 +43,7 @@ void gdr::albedo_pass::Initialize(void)
 
         {
             CD3DX12_ROOT_SIGNATURE_DESC rootSignatureDesc;
-            rootSignatureDesc.Init((UINT)params.size(), &params[0], Render->TexturesSystem->SamplersDescs.size(), Render->TexturesSystem->SamplersDescs.data(), D3D12_ROOT_SIGNATURE_FLAG_ALLOW_INPUT_ASSEMBLER_INPUT_LAYOUT);
+            rootSignatureDesc.Init((UINT)params.size(), &params[0], (UINT)Render->TexturesSystem->SamplersDescs.size(), Render->TexturesSystem->SamplersDescs.data(), D3D12_ROOT_SIGNATURE_FLAG_ALLOW_INPUT_ASSEMBLER_INPUT_LAYOUT);
             Render->GetDevice().CreateRootSignature(rootSignatureDesc, &RootSignature);
         }
     }
